requirements.txt
msal
requests
pyadomd
pydantic
langchain
dotenv
pyyaml
